requirements.txt
scrapy==2.11.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
beautifulsoup4==4.12.2
aiohttp==3.9.1
fake-useragent==1.4.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
validators==0.22.0
